Rename: ChallengeType,GoalType 페키지 변경, GoalType -> ChallengeGoalType으로 변경
