+++ src/main/java/com/dnd/runus/global/constant/ChallengeGoalType.java
@@ -1,4 +1,4 @@
-package com.dnd.runus.infrastructure.persistence.jpa.challenge.entity;
+package com.dnd.runus.global.constant;
 
 /**
  * GoalType은 오늘의 챌린지의 목표 타겟에 대한 챌린지 타입이다.
@@ -7,7 +7,7 @@ package com.dnd.runus.infrastructure.persistence.jpa.challenge.entity;
  * <p> {@code TIME} : 챌린지 목표 타겟이 시간인 타입
  * <p> {@code PACE} : 챌린지 목표 타겟이 페이스인 타입
  */
-public enum GoalType {
+public enum ChallengeGoalType {
     DISTANCE,
     TIME,
     PACE;
